Move CustomWidget and Renderer to the common samples because that already had a dependency on skia and makes the sample easier to work with
