+++ Mapsui.sln
@@ -101,8 +101,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "api", "api", "{5BB9C8BA-0FF
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mapsui.Samples.Droid", "Samples\Mapsui.Samples.Droid\Mapsui.Samples.Droid.csproj", "{3BC0AC7E-3B3B-4D3F-9CC0-B2BDECC460A6}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mapsui.Samples.CustomWidget", "Samples\Mapsui.Samples.CustomWidget\Mapsui.Samples.CustomWidget.csproj", "{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mapsui.UI.WinUI", "Mapsui.UI.WinUI\Mapsui.UI.WinUI.csproj", "{27B81FBB-F6E9-49BB-BAE9-F99F0E01DB73}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mapsui.Samples.WinUI", "Samples\Mapsui.Samples.WinUI\Mapsui.Samples.WinUI\Mapsui.Samples.WinUI.csproj", "{F8B5F48E-5412-4D7A-A76F-01198D6E1768}"
@@ -953,62 +951,6 @@ Global
 		{3BC0AC7E-3B3B-4D3F-9CC0-B2BDECC460A6}.Release|x86.ActiveCfg = Release|Any CPU
 		{3BC0AC7E-3B3B-4D3F-9CC0-B2BDECC460A6}.Release|x86.Build.0 = Release|Any CPU
 		{3BC0AC7E-3B3B-4D3F-9CC0-B2BDECC460A6}.Release|x86.Deploy.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|AnyCPU.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|AnyCPU.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|ARM.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|ARM64.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|ARM64.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|iPhone.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|iPhone.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|x64.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|x64.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|x86.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Ad-Hoc|x86.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|AnyCPU.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|AnyCPU.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|ARM.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|ARM.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|ARM64.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|ARM64.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|iPhone.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|iPhone.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|x64.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|x64.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|x86.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.AppStore|x86.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|AnyCPU.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|AnyCPU.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|ARM.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|x64.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Debug|x86.Build.0 = Debug|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|AnyCPU.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|AnyCPU.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|ARM.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|ARM.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|ARM64.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|iPhone.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|x64.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|x64.Build.0 = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|x86.ActiveCfg = Release|Any CPU
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044}.Release|x86.Build.0 = Release|Any CPU
 		{27B81FBB-F6E9-49BB-BAE9-F99F0E01DB73}.Ad-Hoc|AnyCPU.ActiveCfg = Debug|Any CPU
 		{27B81FBB-F6E9-49BB-BAE9-F99F0E01DB73}.Ad-Hoc|AnyCPU.Build.0 = Debug|Any CPU
 		{27B81FBB-F6E9-49BB-BAE9-F99F0E01DB73}.Ad-Hoc|ARM.ActiveCfg = Debug|Any CPU
@@ -2853,7 +2795,6 @@ Global
 		{52C84C7C-1D60-454E-804A-1F9691BD61A3} = {7D976D37-0542-4A0D-99E6-6BAA180E7011}
 		{5BB9C8BA-0FF4-40E7-BC36-8D671A951C82} = {97458DE1-C297-4DED-AECB-5FEFED2B5610}
 		{3BC0AC7E-3B3B-4D3F-9CC0-B2BDECC460A6} = {969DB691-49FB-4785-98F4-2112CC1BA7C4}
-		{3EB5D6AA-8B52-4E93-9000-61EEE6D6B044} = {969DB691-49FB-4785-98F4-2112CC1BA7C4}
 		{F8B5F48E-5412-4D7A-A76F-01198D6E1768} = {969DB691-49FB-4785-98F4-2112CC1BA7C4}
 		{A1FFEB6C-BE22-4659-83D2-F80333A6E7FB} = {F7130D3D-4E62-4EB3-B74C-1E819E017048}
 		{4EA85C26-FD2A-41A2-910A-23057CAF6616} = {F7130D3D-4E62-4EB3-B74C-1E819E017048}
