+++ Samples/Mapsui.Samples.Common/Maps/Widgets/CustomWidgetSample.cs
@@ -1,7 +1,13 @@
 ﻿using Mapsui.Extensions;
+using Mapsui.Manipulations;
+using Mapsui.Rendering.Skia.SkiaWidgets;
 using Mapsui.Styles;
 using Mapsui.Tiling;
+using Mapsui.Widgets;
+using SkiaSharp;
+using System;
 using System.Threading.Tasks;
+using Mapsui.Rendering.Skia.Extensions;
 
 namespace Mapsui.Samples.Common.Maps.Widgets;
 
@@ -16,7 +22,7 @@ public class CustomWidgetSample : ISample
         var map = new Map();
 
         map.Layers.Add(OpenStreetMap.CreateTileLayer());
-        map.Widgets.Add(new CustomWidget.CustomWidget
+        map.Widgets.Add(new CustomWidget
         {
             Margin = new MRect(20),
             Width = 100,
@@ -25,5 +31,59 @@ public class CustomWidgetSample : ISample
         });
 
         return Task.FromResult(map);
+    }
+
+    public class CustomWidget : BaseWidget
+    {
+        private static readonly Random _random = new();
+
+        public CustomWidget()
+        {
+            Margin = new(20);
+        }
+
+        public Color? Color { get; set; }
+
+        public override bool OnTapped(Navigator navigator, WidgetEventArgs e)
+        {
+            base.OnTapped(navigator, e);
+
+            if (e.TapType == TapType.Single)
+                Color = GenerateRandomColor();
+            else
+                Color = Color.Transparent;
+            return false;
+        }
+
+        public static Color GenerateRandomColor()
+        {
+            byte[] rgb = new byte[3];
+            _random.NextBytes(rgb);
+            return new Color(rgb[0], rgb[1], rgb[2]);
+        }
+    }
+
+    public class CustomWidgetSkiaRenderer : ISkiaWidgetRenderer
+    {
+        public void Draw(SKCanvas canvas, Viewport viewport, IWidget widget, float layerOpacity)
+        {
+            // Cast to custom widget to be able to access the specific CustomWidget fields
+            var customWidget = (CustomWidget)widget;
+
+            // Update the envelope so the MapControl can do hit detection
+            widget.Envelope = ToEnvelope(customWidget);
+
+            // Use the envelope to draw
+            using var skPaint = new SKPaint { Color = customWidget.Color.ToSkia(layerOpacity) };
+            canvas.DrawRect(widget.Envelope.ToSkia(), skPaint);
+        }
+
+        private static MRect ToEnvelope(CustomWidget customWidget)
+        {
+            // A better implementation would take into account widget alignment
+            return new MRect(customWidget.Margin.Left, customWidget.Margin.Top,
+                customWidget.Margin.Left + customWidget.Width,
+                customWidget.Margin.Top + customWidget.Height);
+        }
     }
 }
