+++ Samples/Mapsui.Samples.Maui.MapView/Mapsui.Samples.Maui.MapView.csproj
@@ -81,14 +81,11 @@
 		<ProjectReference Include="..\..\Mapsui\Mapsui.csproj" />
 		<ProjectReference Include="..\..\Tests\Mapsui.Tests.Common\Mapsui.Tests.Common.csproj" />
 		<ProjectReference Include="..\Mapsui.Samples.Common\Mapsui.Samples.Common.csproj" />
-		<ProjectReference Include="..\Mapsui.Samples.CustomWidget\Mapsui.Samples.CustomWidget.csproj" />
 	</ItemGroup>
 
   <!--Source Code Generators-->
 	<ItemGroup>
-		<ProjectReference Include="..\..\SourceGenerators\Mapsui.Sample.SourceGenerator\Mapsui.Sample.SourceGenerator.csproj"
-                      OutputItemType="Analyzer"
-                      ReferenceOutputAssembly="false"  />
+		<ProjectReference Include="..\..\SourceGenerators\Mapsui.Sample.SourceGenerator\Mapsui.Sample.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
 	</ItemGroup>
 
 	<Import Project="..\Mapsui.Samples.MapView\Mapsui.Samples.MapView.projitems" Label="Shared" />
