+++ Samples/Mapsui.Samples.Maui/Mapsui.Samples.Maui.csproj
@@ -64,7 +64,6 @@
 		<ProjectReference Include="..\..\Mapsui.UI.Maui\Mapsui.UI.Maui.csproj" />
 		<ProjectReference Include="..\..\Tests\Mapsui.Tests.Common\Mapsui.Tests.Common.csproj" />
 		<ProjectReference Include="..\Mapsui.Samples.Common\Mapsui.Samples.Common.csproj" />
-		<ProjectReference Include="..\Mapsui.Samples.CustomWidget\Mapsui.Samples.CustomWidget.csproj" />
 	</ItemGroup>
 
 </Project>
