+++ Samples/Mapsui.Samples.Uno.WinUI/Mapsui.Samples.Uno.WinUI.Wasm/Mapsui.Samples.Uno.WinUI.Wasm.csproj
@@ -91,7 +91,6 @@
 		<ProjectReference Include="..\..\..\Mapsui.UI.Uno.WinUI\Mapsui.UI.Uno.WinUI.csproj" />
 		<ProjectReference Include="..\..\..\Tests\Mapsui.Tests.Common\Mapsui.Tests.Common.csproj" />
 		<ProjectReference Include="..\..\Mapsui.Samples.Common\Mapsui.Samples.Common.csproj" />
-		<ProjectReference Include="..\..\Mapsui.Samples.CustomWidget\Mapsui.Samples.CustomWidget.csproj" />
 		<ProjectReference Include="..\Mapsui.Samples.Uno.WinUI\Mapsui.Samples.Uno.WinUI.csproj" />
 	</ItemGroup>
 	<Import Project="..\Mapsui.Samples.Uno.WinUI.Base\base.props" />
